# server.js
const express = require('express');
// const cors = require('cors');
const db = require('./db/bikes.js');
const swaggerUi = require("swagger-ui-express");
const swaggerDocument = require("./swagger.json");
// const swaggerJsdoc = require("swagger-jsdoc");
// const swaggerSpec =swaggerJsdoc(swaggerDocument);
const app = express();
// app.use(cors());
app.use(express.json());
app.use(express.urlencoded({ extended: true }));
app.get('/bikes', async (req, res) => {
  const getBikes = await db.getAllBikes();
  res.json({ getBikes });
});

app.post('/bikes', async (req, res) => {
  const id = await db.createBike(req.body);
  const postDetails = req.body;
  res.json({ success: true, data: postDetails });
});
app.get('/bikes/:id', async (req, res) => {
  const getBike = await db.getBike(req.params.id);
  res.json({ data: getBike });
});
app.patch('/bikes/:id', async (req, res) => {
  const bikeUpdate = await db.updateBike(req.params.id, req.body);
  res.json({ success: `${'updated bike with id number'} ${bikeUpdate}` });
});

app.delete('/bikes/:id', async (req, res) => {
  const bikeDelete = await db.deleteBike(req.params.id);
  res.json({ success: 'bike with id' + bikeDelete +  'is deleted' });
});

app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.listen(3000 || process.env.PORT, () => {
  console.log('server is running on port number 3000');
});